requirements.txt
streamlit
google-generativeai
pymongo
bcrypt
python-dotenv
numpy
matplotlib
dnspython
